feat(data-workflows): Provided item view data for conf itemViews
