+++ src/mixins/data-workflows/doc/doc-mixin.ts
@@ -4,12 +4,43 @@
  */
 
 import { Component, Vue, Watch } from 'vue-property-decorator';
-import { AnyObject, DataWorkflowsType } from '@/types';
+import { AnyObject, DataWorkflowsType, Tab } from '@/types';
 import { IPluginState } from 'vuex-easy-firestore/types/declarations';
 import DataManagementHeader from '@/components/data-workflows/common/DataManagementHeader.vue';
 import ItemComponent from '@/components/data-workflows/common/item/ItemComponent.vue';
 import store from '@/store';
 import { mapState } from 'vuex';
+
+export interface DataItem {
+	component: string;
+	props: AnyObject;
+}
+
+interface ConfigurationTab extends Tab {
+	component: {
+		name: string;
+		props: {
+			data: DataItem[];
+		};
+	};
+}
+
+interface FullJSONTab extends Tab {
+	component: {
+		name: string;
+		props: {
+			json: AnyObject;
+			jsonId: string;
+		};
+	};
+}
+
+interface ConversationTab extends Tab {
+	component: {
+		name: string;
+		props: {};
+	};
+}
 
 @Component({
 	components: { DataManagementHeader, ItemComponent },
@@ -25,6 +56,7 @@ export default class DocMixin extends Vue {
 	private firestoreItem: any;
 	archivedConfsModuleName: any;
 	itemTabsItems: any;
+	configurationData!: DataItem[];
 
 	// TODO: See if this watch decorator is removable
 	@Watch('firestoreItem')
@@ -63,5 +95,43 @@ export default class DocMixin extends Vue {
 	// Overridden by component
 	get type(): DataWorkflowsType {
 		return null;
+	}
+
+	get configurationTab(): ConfigurationTab {
+		return {
+			label: 'Configuration',
+			href: 'configuration',
+			component: {
+				name: 'overview-component',
+				props: {
+					data: this.configurationData
+				}
+			}
+		};
+	}
+
+	get fullJSONTab(): FullJSONTab {
+		return {
+			label: 'Full Json',
+			href: 'full-json',
+			component: {
+				name: 'view-json',
+				props: {
+					json: this.item,
+					jsonId: this.itemId
+				}
+			}
+		};
+	}
+
+	get conversationTab(): ConversationTab {
+		return {
+			label: 'Conversation',
+			href: 'conversation',
+			component: {
+				name: 'view-conversation',
+				props: {}
+			}
+		};
 	}
 }
