+++ src/views/data-workflows/gcs-to-gcs/configurations/ItemView.vue
@@ -19,7 +19,6 @@ export default class GcsToGcsConfigurationsItemView extends Mixins(HeaderInfosMi
 	moduleName: string = mirrorExcGcsToGcsConfs.moduleName;
 	archivedConfsModuleName: string = mirrorExcGcsToGcsConfsArchive.moduleName;
 
-	// TODO: Move to dedicated file all methods / computed below
 	getDestinationStorageRows() {
 		let destinationStorageRows = [];
 
@@ -41,42 +40,6 @@ export default class GcsToGcsConfigurationsItemView extends Mixins(HeaderInfosMi
 				filename_description: 'No Description'
 			};
 		});
-	}
-
-	get itemTabsItems() {
-		if (Object.keys(this.item).length === 0) return [];
-
-		return [
-			{
-				label: 'Configuration',
-				href: 'configuration',
-				component: {
-					name: 'overview-component',
-					props: {
-						data: this.configurationData
-					}
-				}
-			},
-			{
-				label: 'Full Json',
-				href: 'full-json',
-				component: {
-					name: 'view-json',
-					props: {
-						json: this.item,
-						jsonId: this.itemId
-					}
-				}
-			},
-			{
-				label: 'Conversation',
-				href: 'conversation',
-				component: {
-					name: 'view-conversation',
-					props: {}
-				}
-			}
-		];
 	}
 
 	get configurationData() {
