+++ src/views/data-workflows/storage-to-storage/configurations/ItemView.vue
@@ -20,7 +20,6 @@ export default class StorageToStorageConfigurationsItemView extends Mixins(Heade
 	moduleName: string = storageToStorageConfs.moduleName;
 	archivedConfsModuleName: string = storageToStorageConfsArchive.moduleName;
 
-	// TODO: Move to dedicated file all methods / computed below
 	getSourceStorageColumns() {
 		let sourceStorageColumns = {};
 
@@ -140,42 +139,6 @@ export default class StorageToStorageConfigurationsItemView extends Mixins(Heade
 		destinationStorageRows.s3 = s3DestinationStorageRows;
 		destinationStorageRows.sftp = sftpDestinationStorageRows;
 		return destinationStorageRows;
-	}
-
-	get itemTabsItems() {
-		if (Object.keys(this.item).length === 0) return [];
-
-		return [
-			{
-				label: 'Configuration',
-				href: 'configuration',
-				component: {
-					name: 'overview-component',
-					props: {
-						data: this.configurationData
-					}
-				}
-			},
-			{
-				label: 'Full Json',
-				href: 'full-json',
-				component: {
-					name: 'view-json',
-					props: {
-						json: this.item,
-						jsonId: this.itemId
-					}
-				}
-			},
-			{
-				label: 'Conversation',
-				href: 'conversation',
-				component: {
-					name: 'view-conversation',
-					props: {}
-				}
-			}
-		];
 	}
 
 	get configurationData() {
